requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
macholib==1.16.3
oauthlib==3.2.2
openai==1.76.0
packaging==25.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.0.1
sniffio==1.3.1
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
tk==0.1.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.30.0
